+++ example/android/app/src/main/kotlin/cn/chessroad/anco_env_example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.anco_env_example
+package cn.chessroad.anco_env_example
 
 import io.flutter.embedding.android.FlutterActivity