+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.0.1
+## 0.1.0-dev.1
 
-* TODO: Describe initial release.
+* Initial release.
+* Added `isHarmonyAnco` to detect if the app is running in the HarmonyOS Anco environment.
+* Added `getEnvInfo` to collect comprehensive system environment information.
+* Added `getProp` to retrieve system properties on Android.